refactor(test): introduce --stack-config and simplify options
